pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "old-bailey-mcp-v2"
version = "2.0.0"
description = "Old Bailey Online MCP server — living research memory for 240 years of London criminal court records"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.2.3",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=9.0.0"]

[tool.hatch.build.targets.wheel]
packages = []
